requirements.txt
# Core AI and ML dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0

# Voice processing and cloning
TTS>=0.14.0
coqui-ai-tts>=0.14.0
speechbrain>=0.5.16
librosa>=0.10.1
soundfile>=0.12.1
pyaudio>=0.2.11
webrtcvad>=2.0.10
openai-whisper>=20231117

# Computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
mediapipe>=0.10.0
ultralytics>=8.0.0

# Audio processing
pydub>=0.25.1
scipy>=1.11.0
matplotlib>=3.7.0

# Web framework and UI
streamlit>=1.28.0
gradio>=4.0.0
flask>=2.3.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: GPU acceleration
# cuda-python>=12.0.0  # Uncomment if using CUDA 